For NMS, boxes typically have 5 or 6 features which are the 4 box coordinates, per-box scores, and sometimes per-box classes. However, the boxes are also allowed to have any amount of additional features. We didn't have any unit tests for that situation, so have added one. After recent changes to CUDA nms implementation, those additional features were not being copied around anymore.

Additional features per box: https://mxnet.incubator.apache.org/versions/1.7.0/api/python/docs/api/symbol/contrib/index.html#mxnet.symbol.contrib.box_nms

> By default, a box is [id, score, xmin, ymin, xmax, ymax, …], additional elements are allowed.

@masahi @mbrookhart @anijain2305