@masahi 2021-02-20
please verify that there is no perf regression for normal cases (PyTorch, TF). I now understand why MXNet NMS expects weird packed inputs lol