@trevor-m 2021-02-22
@masahi Thank you for reviewing! The performance seems to be the same with same random variance.

| Frontend | #Boxes | Without this PR | With this PR |
|---|---|---|---|
| Pytorch | 64 | 0.05 ms | 0.04 ms |
| Pytorch | 1024  | 1.10 ms | 1.09 ms |
| Pytorch | 4500 | 44.73 ms | 44.59 ms |
| TensorFlow | 64 | 0.05 ms | 0.05 ms |
| TensorFlow | 1024 | 0.11 ms | 0.10 ms |
| TensorFlow | 67500 | 4.91 ms | 4.95 ms |