@masahi 2021-02-22
@trevor-m Thanks, btw is the PyTorch 4500 boxes number coming from a real workload? It seems really slow. If it is multiclass NMS, I suggest using graph rewrite I added for it (see `pytorch/test_object_detection.py`)