@trevor-m 2021-02-22
> @trevor-m Thanks, btw is the PyTorch 4500 boxes number coming from a real workload? It seems really slow. If it is multiclass NMS, I suggest using graph rewrite I added for it (see `pytorch/test_object_detection.py`)

No it's not real, I just edited the box count for the Pytorch unit test `test_forward_nms`